+++ javaparser-symbol-solver-core/src/main/java/com/github/javaparser/symbolsolver/javaparsermodel/contexts/AnonymousClassDeclarationContext.java
@@ -41,8 +41,7 @@ public class AnonymousClassDeclarationContext extends AbstractJavaParserContext<
   @Override
   public SymbolReference<ResolvedMethodDeclaration> solveMethod(String name,
                                                                 List<ResolvedType> argumentsTypes,
-                                                                boolean staticOnly,
-                                                                TypeSolver typeSolver) {
+                                                                boolean staticOnly) {
     List<ResolvedMethodDeclaration> candidateMethods =
         myDeclaration
             .getDeclaredMethods()
@@ -56,8 +55,7 @@ public class AnonymousClassDeclarationContext extends AbstractJavaParserContext<
             MethodResolutionLogic.solveMethodInType(ancestor.getTypeDeclaration(),
                                                     name,
                                                     argumentsTypes,
-                                                    staticOnly,
-                                                    typeSolver);
+                                                    staticOnly);
         // consider methods from superclasses and only default methods from interfaces :
         // not true, we should keep abstract as a valid candidate
         // abstract are removed in MethodResolutionLogic.isApplicable is necessary
@@ -71,7 +69,7 @@ public class AnonymousClassDeclarationContext extends AbstractJavaParserContext<
     // see issue #75
     if (candidateMethods.isEmpty()) {
       SymbolReference<ResolvedMethodDeclaration> parentSolution =
-          getParent().solveMethod(name, argumentsTypes, staticOnly, typeSolver);
+          getParent().solveMethod(name, argumentsTypes, staticOnly);
       if (parentSolution.isSolved()) {
         candidateMethods.add(parentSolution.getCorrespondingDeclaration());
       }
@@ -84,8 +82,7 @@ public class AnonymousClassDeclarationContext extends AbstractJavaParserContext<
                                                                                  typeSolver),
                                                   name,
                                                   argumentsTypes,
-                                                  false,
-                                                  typeSolver);
+                                                  false);
       if (res.isSolved()) {
         candidateMethods.add(res.getCorrespondingDeclaration());
       }
@@ -98,7 +95,7 @@ public class AnonymousClassDeclarationContext extends AbstractJavaParserContext<
   }
 
   @Override
-  public SymbolReference<ResolvedTypeDeclaration> solveType(String name, TypeSolver typeSolver) {
+  public SymbolReference<ResolvedTypeDeclaration> solveType(String name) {
     List<com.github.javaparser.ast.body.TypeDeclaration> typeDeclarations =
         myDeclaration
             .findMembersOfKind(com.github.javaparser.ast.body.TypeDeclaration.class);
@@ -124,8 +121,7 @@ public class AnonymousClassDeclarationContext extends AbstractJavaParserContext<
             .map(internalType ->
                      JavaParserFactory
                          .getContext(internalType, typeSolver)
-                         .solveType(name.substring(internalType.getName().getId().length() + 1),
-                                    typeSolver));
+                         .solveType(name.substring(internalType.getName().getId().length() + 1)));
 
     if (recursiveMatch.isPresent()) {
       return recursiveMatch.get();
@@ -168,19 +164,18 @@ public class AnonymousClassDeclarationContext extends AbstractJavaParserContext<
       }
     }
     
-    return getParent().solveType(name, typeSolver);
+    return getParent().solveType(name);
   }
 
   @Override
-  public SymbolReference<? extends ResolvedValueDeclaration> solveSymbol(String name,
-                                                                         TypeSolver typeSolver) {
+  public SymbolReference<? extends ResolvedValueDeclaration> solveSymbol(String name) {
     Preconditions.checkArgument(typeSolver != null);
 
     if (myDeclaration.hasVisibleField(name)) {
       return SymbolReference.solved(myDeclaration.getVisibleField(name));
     }
 
-    return getParent().solveSymbol(name, typeSolver);
+    return getParent().solveSymbol(name);
   }
 
 }