+++ changelog.md
@@ -1,3 +1,10 @@
+Version 3.7.1
+------------------
+* slightly breaking: the enum constants in JsonToken are now capitalized.
+* slightly breaking: [some obscure methods in the symbol solver changed](https://github.com/javaparser/javaparser/pull/1922) 
+
+[issues resolved](https://github.com/javaparser/javaparser/milestone/113?closed=1)
+
 Version 3.7.0
 ------------------
 * BREAKING: `ForeachStmt` is now correctly capitalized: `ForEachStmt`